+++ internal/provider/model/job_definition/output_option/bigquery.go
@@ -2,7 +2,7 @@ package output_options
 
 import (
 	"terraform-provider-trocco/internal/client/entity/job_definition/output_option"
-	output_options2 "terraform-provider-trocco/internal/client/parameter/job_definition/output_options"
+	output_options2 "terraform-provider-trocco/internal/client/parameter/job_definition/output_option"
 	"terraform-provider-trocco/internal/provider/model"
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
